Rename _tf files
